XP bar fixes
Merging 2733736

//depot/UE4-UT/...

to //depot/UE4-UT-Releases/...

[CL 2733803 by Peter Knepley in UT-Release branch]

+++ UnrealTournament/Source/UnrealTournament/Private/Slate/Widgets/SUTXPBar.h
@@ -54,7 +54,7 @@ protected:
 
 	FMargin GetBreakdownMargin() const;
 	TOptional<FSlateRenderTransform> GetLevelTransform() const;
-	int32 GetXP() const;
+	float GetXP() const;
 	FXPBreakdown GetXPBreakdown() const;
 
 	virtual void OnLevelUp(int32 NewLevel);
